Create Day5
This C++ program performs basic arithmetic operations (addition, subtraction, multiplication, division, modulus) based on user input.

Input: The user inputs two integers (a, b) and selects an operator (+, -, *, /, %).

Switch Case: Based on the operator, the program performs the corresponding operation. If an invalid operator is entered, it prints an error message.

Output: The result of the operation is displayed.