Intro to Array
This C++ program searches for a specific element (key) in an array.

search Function: Loops through the array, checking each element. If it finds the key, it returns 1 (true), otherwise returns 0 (false).

Main Function: The user inputs a key to search for in the predefined array. The program calls the search function and outputs whether the key is present or absent in the array.